CIRC-1061 update tests according to code review
