+++ src/main/java/org/folio/circulation/domain/representations/ItemLimitBlock.java
@@ -1,9 +1,6 @@
 
 package org.folio.circulation.domain.representations;
 
-import lombok.Getter;
-
-@Getter
 public class ItemLimitBlock {
 
 }
