+++ src/main/java/org/folio/circulation/support/http/server/InsufficientOverridePermissionsError.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import org.folio.circulation.resources.handlers.error.OverridableBlockType;
 
-public class InsufficientOverridePermissionsError extends OverridableValidationError {
+public class InsufficientOverridePermissionsError extends OverrideValidationError {
 
   public InsufficientOverridePermissionsError(OverridableBlockType blockType,
     List<String> missingOverridePermissions) {
